# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VitaSense AI – symptom analysis and health guidance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /symptoms/analyze:
    post:
      operationId: analyzeSymptoms
      tags: [symptoms]
      summary: Run AI-powered symptom analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SymptomInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /symptoms/suggestions:
    get:
      operationId: getSymptomSuggestions
      tags: [symptoms]
      summary: Get symptom autocomplete suggestions
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of symptom suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SymptomSuggestion"
...
components:
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    SymptomInput:
      type: object
      required: [symptoms, duration, ageRange, severity]
      properties:
        symptoms:
          type: array
          items:
            type: string
          minItems: 1
        duration:
          type: string
          enum: [few_hours, one_day, several_days, one_week_plus, chronic]
        ageRange:
          type: string
          enum: [child, teen, adult, elderly]
        severity:
          type: string
          enum: [mild, moderate, severe, emergency]
        additionalNotes:
          type: ["string", "null"]
    AnalysisResult:
      type: object
      required: [urgencyLevel, urgencyLabel, possibleConditions, lifestyleAdvice, whenToSeeDoctor, disclaimer]
      properties:
        urgencyLevel:
          type: string
          enum: [mild_concern, moderate_concern, seek_medical_attention, emergency_care]
        urgencyLabel:
          type: string
        possibleConditions:
          type: array
          items:
            $ref: "#/components/schemas/PossibleCondition"
        lifestyleAdvice:
          type: array
          items:
            type: string
        whenToSeeDoctor:
          type: array
          items:
            type: string
        disclaimer:
          type: string
        summary:
          type: ["string", "null"]
    PossibleCondition:
      type: object
      required: [name, description, likelihood, commonCauses, riskFactors, basicApproaches]
      properties:
        name:
          type: string
        description:
          type: string
        likelihood:
          type: integer
          description: Percentage likelihood 0-100
        commonCauses:
          type: array
          items:
            type: string
        riskFactors:
          type: array
          items:
            type: string
        basicApproaches:
          type: array
          items:
            type: string
    SymptomSuggestion:
      type: object
      required: [id, label, category]
      properties:
        id:
          type: string
        label:
          type: string
        category:
          type: string
...