# lib/api-zod/src/generated/types/possibleCondition.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VitaSense AI – symptom analysis and health guidance API
 * OpenAPI spec version: 0.1.0
 */

export interface PossibleCondition {
  name: string;
  description: string;
  /** Percentage likelihood 0-100 */
  likelihood: number;
  commonCauses: string[];
  riskFactors: string[];
  basicApproaches: string[];
}